+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 0.1.0-alpha.8 [Unreleased]
 ### Added
   - #49 Add new filter interface `Fiv\Form\Filter\StringFilter\StringFilterInterface`
+### Removed  
+  - #48 TextArea now implements `StringDataElementInterface`. Find usages and rewrite
     
 ## 0.1.0-alpha.7 [2017-02-01]
 ### Added